Add automatically controlled player 2
The 2nd player (right side of screen) now automatically follows
the ball, instead of requiring keyboard input.
This makes the game playable with a single person.
The automated player only reacts to the ball when the ball is in its half.
I haven't put any effort into balancing the game so it's probably far too hard.